requirements.txt
Pillow>=10.0.0
cryptography>=40.0.0
